Extract mod/riscv64
(cherry picked from commit fa493710a7db004072f43346c5cd5566b6634ae7)
